refactor: break some business logic into lib
