+++ src/main.rs
@@ -1,99 +1,10 @@
 extern crate dotenv;
 
 use dotenv::dotenv;
-use serde::{Deserialize, Serialize};
 use serde_json::Value;
-use std::{env, error::Error, fs::File};
+use std::{env, error::Error};
 
-const MIX_MAX_SEQUENCE_RECIPIENTS_URL: &str = "https://api.mixmax.com/v1/sequences/:id/recipients";
-
-#[derive(Clone, Deserialize, Debug)]
-struct Person {
-    #[serde(rename = "First Name")]
-    first_name: String,
-
-    #[serde(rename = "Last Name")]
-    last_name: String,
-
-    #[serde(rename = "Company Name")]
-    company_name: String,
-
-    #[serde(rename = "Job Title")]
-    job_title: String,
-
-    #[serde(rename = "Email")]
-    email_address: String,
-
-    #[serde(rename = "LinkedIn URL")]
-    linkedin_url: String,
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct RequestBody {
-    recipients: Vec<Recipient>,
-    #[serde(rename = "scheduledAt")]
-    scheduled_at: bool,
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct Recipient {
-    #[serde(rename = "email")]
-    email_address: String,
-    #[serde(rename = "variables")]
-    variables: RecipientVariables,
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct RecipientVariables {
-    #[serde(rename = "First Name")]
-    first_name: String,
-
-    #[serde(rename = "Last Name")]
-    last_name: String,
-
-    #[serde(rename = "Company Name")]
-    company_name: String,
-
-    #[serde(rename = "Job Title")]
-    job_title: String,
-
-    #[serde(rename = "Email")]
-    email_address: String,
-
-    #[serde(rename = "LinkedIn URL")]
-    linkedin_url: String,
-}
-
-impl From<Person> for RecipientVariables {
-    fn from(person: Person) -> Self {
-        RecipientVariables {
-            first_name: person.first_name,
-            last_name: person.last_name,
-            company_name: person.company_name,
-            job_title: person.job_title,
-            email_address: person.email_address,
-            linkedin_url: person.linkedin_url,
-        }
-    }
-}
-
-fn parse_people(csv_reader: &mut csv::Reader<File>) -> Result<Vec<Person>, Box<dyn Error>> {
-    let people: Vec<Person> = csv_reader.deserialize().collect::<Result<_, _>>()?;
-    let people = people
-        .into_iter()
-        .filter(|p| {
-            let cleaned_email = p
-                .email_address
-                .chars()
-                .filter(|&c| c != '\u{2060}')
-                .collect::<String>();
-
-            !cleaned_email.is_empty()
-        })
-        .collect::<Vec<Person>>();
-
-    Ok(people)
-}
+use mixmax_csv_uploader::{parse_people, Recipient, RecipientVariables, RequestBody};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
@@ -119,7 +30,8 @@ async fn main() -> Result<(), Box<dyn Error>> {
     let sequence_id = env::var("MIXMAX_SEQUENCE_ID").expect("MIXMAX Sequence ID is expected");
 
     // Construct URL for Mix Max Sequence receipients
-    let full_url = MIX_MAX_SEQUENCE_RECIPIENTS_URL.replace(":id", &sequence_id);
+    let full_url =
+        mixmax_csv_uploader::MIX_MAX_SEQUENCE_RECIPIENTS_URL.replace(":id", &sequence_id);
     println!("Full URL: {full_url}");
 
     // Send Request
